+++ terraform/modules/helm_postgresql/providers.tf
@@ -3,7 +3,7 @@ terraform {
 
   required_providers {
     helm = {
-      source = "hashicorp/helm"
+      source  = "hashicorp/helm"
       version = "3.0.2"
     }
     random = {