+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Exceptions are acceptable depending on the circumstances (critical bug fixes tha
 
 ### Added
 
+- added a new module to deploy the Vehicle Payment application
+- added a new module to deploy the Vehicle Management application
 - added a new step to plan and deploy the infrastructure
 - added a new module to create the resource group and kubernetes in azure environment
 - created the base structure